requirements.txt
flask
groq
pandas
openpyxl
python-dotenv
